move BinTreeAug into bin_tree_aug mod
