+++ crates/tlb-ton/src/lib.rs
@@ -1,4 +1,5 @@
 mod address;
+pub mod bin_tree_aug;
 pub mod bin_tree;
 mod boc;
 pub mod currency;
@@ -7,6 +8,5 @@ mod message;
 mod state_init;
 mod timestamp;
 mod unary;
-mod bintree;
 
-pub use self::{address::*, boc::*, message::*, state_init::*, timestamp::*, unary::*, bintree::*,};
+pub use self::{address::*, boc::*, message::*, state_init::*, timestamp::*, unary::*,};
